backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
python-multipart==0.0.20
ortools==9.15.6755
sqlalchemy==2.0.40
alembic==1.15.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.3.1